@@ -1,14 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
         <PackageReference Include="Google.Apis.YouTube.v3" Version="1.52.0.2378" />
         <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="5.0.8" />
         <PackageReference Include="Microsoft.AspNetCore.WebUtilities" Version="2.2.0" />
-        <PackageReference Include="Discord.Net" Version="3.7.1" />
     </ItemGroup>
 
     <ItemGroup>
